feat(base page): move all manager content to base page
